// clients/js-kit/src/generated/types/baseExternalPluginAdapterKey.ts
export type BaseExternalPluginAdapterKey =
  | { __kind: 'LifecycleHook'; fields: readonly [Address] }
  | { __kind: 'Oracle'; fields: readonly [Address] }
  | { __kind: 'AppData'; fields: readonly [BasePluginAuthority] }
  | { __kind: 'LinkedLifecycleHook'; fields: readonly [Address] }
  | { __kind: 'LinkedAppData'; fields: readonly [BasePluginAuthority] }
  | { __kind: 'DataSection'; fields: readonly [BaseLinkedDataKey] };